fix(ios): reserve liquid glass for controls
