--- platforms/ios/Funput/Components/AdaptiveGlassCard.swift
@@ -1,33 +1,0 @@
-import SwiftUI
-
-struct AdaptiveGlassCard<Content: View>: View {
-    private let content: Content
-    private let cornerRadius: CGFloat
-
-    init(cornerRadius: CGFloat = 22, @ViewBuilder content: () -> Content) {
-        self.cornerRadius = cornerRadius
-        self.content = content()
-    }
-
-    var body: some View {
-        if #available(iOS 26, *) {
-            cardContent
-                .padding(18)
-                .glassEffect(.regular, in: .rect(cornerRadius: cornerRadius))
-        } else {
-            cardContent
-                .padding(18)
-                .background(
-                    .regularMaterial,
-                    in: RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
-                )
-        }
-    }
-
-    private var cardContent: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            content
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
