+++ platforms/ios/Funput/Components/ContentCard.swift
@@ -1,0 +1,29 @@
+import SwiftUI
+
+/// A passive content surface. Liquid Glass is reserved for navigation and
+/// interactive controls so scrolling and tab transitions stay inexpensive.
+struct ContentCard<Content: View>: View {
+    private let content: Content
+    private let cornerRadius: CGFloat
+
+    init(cornerRadius: CGFloat = 22, @ViewBuilder content: () -> Content) {
+        self.cornerRadius = cornerRadius
+        self.content = content()
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 12) {
+            content
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .padding(18)
+        .background(
+            Color(uiColor: .secondarySystemGroupedBackground),
+            in: RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+        )
+        .overlay {
+            RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+                .stroke(Color.primary.opacity(0.07), lineWidth: 0.5)
+        }
+    }
+}
